ImageConverter 별도 파일 생성 및 불필요한 메서드 삭제
