--- Task2.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,23 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "alamofire",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/Alamofire/Alamofire",
-      "state" : {
-        "revision" : "3dc6a42c7727c49bf26508e29b0a0b35f9c7e1ad",
-        "version" : "5.8.1"
-      }
-    },
-    {
-      "identity" : "gzip",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/nicklockwood/GZIP",
-      "state" : {
-        "revision" : "bd55f1d89e71ae3f481da74cd4adeadbb849620c",
-        "version" : "1.3.1"
-      }
-    }
-  ],
-  "version" : 2
-}
