+++ Task2/Views/SeriesTableViewCell.swift
@@ -66,73 +66,3 @@ final class SeriesTableViewCell: UITableViewCell {
         self.dicomImageView.image = model.images[index]
     }
 }
-
-//TODO: Create an individual file for this class.
-final class ImageManger {
-    
-    // [[UInt16]]을 UIImage?로 변환.
-    static func convertUInt16ToImage(from pixelData: [[UInt16]]) -> UIImage? {
-        let data = NSMutableData()
-
-        for row in pixelData {
-            var rowData = row
-            data.append(&rowData, length: MemoryLayout<UInt16>.size * row.count)
-        }
-
-        guard let provider = CGDataProvider(data: data) else {
-            return nil
-        }
-
-        guard let cgImage = CGImage(
-            width: pixelData[0].count,
-            height: pixelData.count,
-            bitsPerComponent: 16,
-            bitsPerPixel: 16,
-            bytesPerRow: MemoryLayout<UInt16>.size * pixelData[0].count,
-            space: CGColorSpaceCreateDeviceGray(),
-            bitmapInfo: CGBitmapInfo(rawValue: CGImageAlphaInfo.none.rawValue),
-            provider: provider,
-            decode: nil,
-            shouldInterpolate: false,
-            intent: .defaultIntent
-        ) else {
-            return nil
-        }
-
-        let image = UIImage(cgImage: cgImage)
-        return image
-    }
-
-    // [[UInt18]]을 UIImage?로 변환.
-    static func convertUInt8ToImage(from pixelData: [[UInt8]]) -> UIImage? {
-        let data = NSMutableData()
-
-        // 2D array를 1D array로 변경.
-        pixelData.forEach {
-            data.append(Data($0))
-        }
-        
-        guard let provider = CGDataProvider(data: data) else {
-            return nil
-        }
- 
-        guard let cgImage = CGImage(
-            width: pixelData[0].count,
-            height: pixelData.count,
-            bitsPerComponent: 8,
-            bitsPerPixel: 8,
-            bytesPerRow: pixelData[0].count,
-            space: CGColorSpaceCreateDeviceGray(),
-            bitmapInfo: .byteOrderDefault,
-            provider: provider,
-            decode: nil,
-            shouldInterpolate: false,
-            intent: .defaultIntent
-        ) else {
-            return nil
-        }
-
-        let image = UIImage(cgImage: cgImage)
-        return image
-    }
-}
